+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fixed `TypeError` when trying to pass composed runtime object to API that
   requires subclass ([#25]).
 - Fixed `TypeError` when passing `None` to APIs that require a `system.Object`.
+- Fixed `ui.xaml.Application.start()` methods not releasing GIL.
 
 [#25]: https://github.com/pywinrt/pywinrt/issues/25